UI and Feature Enhancements
- Completely revamped the user interface across the entire application
- Implemented a markdown editor for creating and editing tickets/answers
- Conducted extensive refactoring to improve code quality and maintainability
- Enhanced existing features and polished the entire application

+++ app/Http/Controllers/AnswerController.php
@@ -24,18 +24,31 @@
         public function __construct( AnswerService $service )
         {
             $this->service = $service;
-            $this->authorizeResource( Answer::class, 'answer' );
+        }
+
+        public function create( Ticket $ticket )
+        {
+            $this->authorize( 'answer', $ticket );
+            return view( 'tickets.answers.create', compact( 'ticket' ) );
         }
 
         public function store( StoreAnswerRequest $request, Ticket $ticket ) : RedirectResponse
         {
+            $this->authorize( 'answer', $ticket );
             $answer = $this->service->create( array_merge( $request->validated(), [ 'ticket_id' => $ticket->id ] ) );
             return redirect()->route( 'tickets.show', $ticket )
                 ->with( 'success', __( 'answers.submitted_successfully' ) );
         }
 
+        public function edit( Answer $answer )
+        {
+            $this->authorize( 'update', $answer );
+            return view( 'tickets.answers.edit', compact( 'answer'  ) );
+        }
+
         public function update( UpdateAnswerRequest $request, Answer $answer ) : RedirectResponse
         {
+            $this->authorize( 'update', $answer );
             $updatedAnswer = $this->service->update( $answer->id, $request->validated() );
             return redirect()->route( 'tickets.show', $answer->ticket )
                 ->with( 'success', __( 'answers.updated_successfully' ) );
@@ -43,6 +56,7 @@
 
         public function destroy( Answer $answer ) : RedirectResponse
         {
+            $this->authorize( 'delete', $answer );
             $ticket = $answer->ticket;
             $this->service->delete( $answer->id );
             return redirect()->route( 'tickets.show', $ticket )
@@ -51,6 +65,7 @@
 
         public function accept( Answer $answer ) : RedirectResponse
         {
+            $this->authorize( 'accept', $answer );
             $this->service->acceptAnswer( $answer );
             return redirect()->route( 'tickets.show', $answer->ticket )
                 ->with( 'success', __( 'answers.accepted_successfully' ) );
